separate build optimization
